Refactor APIs into separate modules
